Move TimeSeriesMappingException to exception package
