+++ metrix-mapping/src/main/groovy/com/powsybl/metrix/mapping/EquipmentMappingData.groovy
@@ -8,6 +8,7 @@
 package com.powsybl.metrix.mapping
 
 import com.powsybl.iidm.network.Identifiable
+import com.powsybl.metrix.mapping.exception.TimeSeriesMappingException
 import groovy.transform.CompileStatic
 
 /**
