+++ metrix-mapping/src/main/java/com/powsybl/metrix/mapping/MappingKeyNetworkValue.java
@@ -9,6 +9,7 @@ package com.powsybl.metrix.mapping;
 
 import com.powsybl.iidm.network.*;
 import com.powsybl.iidm.network.extensions.LoadDetail;
+import com.powsybl.metrix.mapping.exception.TimeSeriesMappingException;
 
 import java.util.Objects;
