+++ metrix-mapping/src/test/java/com/powsybl/metrix/mapping/MappingKeyNetworkValueTest.java
@@ -10,6 +10,7 @@ package com.powsybl.metrix.mapping;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.PhaseTapChanger;
 import com.powsybl.iidm.serde.NetworkSerDe;
+import com.powsybl.metrix.mapping.exception.TimeSeriesMappingException;
 import org.assertj.core.data.Offset;
 import org.junit.jupiter.api.Test;
