+++ metrix-mapping/src/test/java/com/powsybl/metrix/mapping/TimeSeriesDslLoaderTest.java
@@ -13,6 +13,7 @@ import com.google.common.jimfs.Jimfs;
 import com.powsybl.commons.io.table.TableFormatterConfig;
 import com.powsybl.commons.test.TestUtil;
 import com.powsybl.iidm.network.Network;
+import com.powsybl.metrix.mapping.exception.TimeSeriesMappingException;
 import com.powsybl.metrix.mapping.util.MappingTestNetwork;
 import com.powsybl.timeseries.ReadOnlyTimeSeriesStore;
 import com.powsybl.timeseries.ReadOnlyTimeSeriesStoreCache;
